Updated loops

+++ Inventory.cpp
@@ -9,7 +9,7 @@ void Inventory::add_product(Product prdct, int qty) { //function to add product/
         cout<<"Invalid quantity";
         return;
     }
-    for (int i=0;i<products.size();i++) {
+    for (size_t i=0;i<products.size();i++) {
         if (products[i].getProductID()==prdct.getProductID()){
             quantities[i]+=qty;
             return;
@@ -23,7 +23,7 @@ void Inventory::reduce_stock(int id,int amount) { //function to reduce product a
         cout<<"Invalid amount"; //exception handling for purchase amount
         return;
     }
-    for (int i=0;i<products.size();i++) {
+    for (size_t i=0;i<products.size();i++) {
         if (products[i].getProductID()==id) {
             if (quantities[i]>=amount){
                 quantities[i]-=amount;
@@ -39,13 +39,13 @@ void Inventory::reduce_stock(int id,int amount) { //function to reduce product a
 void Inventory::display_inventory() { //function to display all stored products and quantites
     cout << endl;
     cout << "==== Inventory ====" << endl;
-    for (int i=0; i<products.size();i++) {
+    for (size_t i=0; i<products.size();i++) {
         cout<<products[i].getProductName()<<" "<<products[i].getProductID()<<" "<<quantities[i]<<endl;
     }
 }
 // Function to findProduct
 Product* Inventory::findProduct (int ProductID) { 
-    for (int i=0; i<products.size() ;i++) {
+    for (size_t i=0; i<products.size() ;i++) {
         if (products[i].getProductID() == ProductID) {
             return &products[i];
         }
@@ -60,7 +60,7 @@ bool Inventory::updateStock(int productID, int newQuantity) {
         cout<<"Invalid stock quantity"<<endl;
         return false;
     }
-    for (int i=0; i<products.size() ; i++) {
+    for (size_t i=0; i<products.size() ; i++) {
         if (products[i].getProductID() == productID) {
             quantities[i] = newQuantity;
             cout<<"Stock updated successfully"<<endl;
@@ -73,7 +73,7 @@ bool Inventory::updateStock(int productID, int newQuantity) {
 
 // Remove a product
 bool Inventory::removeProduct(int productID) {
-    for (int i=0; i<products.size() ; i++) {
+    for (size_t i=0; i<products.size() ; i++) {
         if (products[i].getProductID()== productID) {
             products.erase(products.begin()+i);
             quantities.erase(quantities.begin()+i);
